Updated conftest to import logically grouped fixture.
