+++ tests/fixtures/__ini__.py
@@ -1,0 +1,5 @@
+"""
+Package initialization file for fixtures.
+
+Makes the fixtures importable when importing from tests.fixtures.
+"""
